+++ axum-macros/tests/typed_path/fail/wildcard.rs
@@ -1,7 +1,0 @@
-use axum_extra::routing::TypedPath;
-
-#[derive(TypedPath)]
-#[typed_path("/users/*rest")]
-struct MyPath;
-
-fn main() {}