+++ axum-macros/CHANGELOG.md
@@ -8,8 +8,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 # Unreleased
 
 - **fixed:** `Option` and `Result` are now supported in typed path route handler parameters ([#1001])
+- **fixed:** Support wildcards in typed paths ([#1003])
 
 [#1001]: https://github.com/tokio-rs/axum/pull/1001
+[#1003]: https://github.com/tokio-rs/axum/pull/1003
 
 # 0.2.0 (31. March, 2022)